Yirmi ikinci Commit

+++ Controllers/HomeController.cs
@@ -168,6 +168,34 @@ namespace FormsApp.Controllers
                 return NotFound();
 
             }
+
+            return View("DeleteConfirm", entity);
+
+        }
+
+
+
+ [HttpPost]
+        public IActionResult Delete(int id, int ProductId)
+
+        {
+
+            if (id != ProductId)
+            {
+                return NotFound();
+            }
+
+            var entity = Repository.Products?.FirstOrDefault(p => p.ProductId == ProductId);
+
+        
+
+            if (entity == null)
+
+            {
+
+                return NotFound();
+
+            }
             Repository.DeleteProduct(entity);
             return RedirectToAction("Index");
